update meeting notes, merger rule checks
Meeting notes: add additional statistics checks on the output. I will also need to add the shellscripts that are currently hosted on the server to the repo, probably....

Merge rules: Added functionality for the minimum-seq-length and maximum-seq length. Also updated the merge rules so that a block MUST contain hg38 to be included (so if the gap is too long for hg38, then the entire block is excluded.) This works for the project I am working on, but perhaps I will want to change it later to generalize, especially to allow for other assemblies such as hg19.